+++ app/Http/Controllers/RemedyController.php
@@ -7,6 +7,24 @@ use Illuminate\Http\Request;
 
 class RemedyController extends Controller
 {
+    public function __construct(Remedy $remedy)
+    {
+        $this->remedy = $remedy;
+    }
+
+    /**
+     * Retrieve specific data from database
+     * 
+     * @param Integer $id
+     */
+    private function find($id) {
+        $remedy = $this->remedy->find($id);
+        if ($remedy === null) {
+            $remedy = ['error' => 'Não foi possível concluir a operação, registro não identificado'];
+        }
+        return $remedy;
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -14,7 +32,7 @@ class RemedyController extends Controller
      */
     public function index()
     {
-        return Remedy::all();
+        return $this->remedy->all();
     }
 
     /**
@@ -25,44 +43,49 @@ class RemedyController extends Controller
      */
     public function store(Request $request)
     {
-        return Remedy::create($request->all());
+        return $this->remedy->create($request->all());
     }
 
     /**
      * Display the specified resource.
      *
-     * @param  \App\Models\Remedy  $remedy
+     * @param  Integer $id
      * @return \Illuminate\Http\Response
      */
     public function show($id)
     {
-        return Remedy::find($id);
+        return $this->find($id);
     }
 
     /**
      * Update the specified resource in storage.
      *
      * @param  \Illuminate\Http\Request  $request
-     * @param  \App\Models\Remedy  $remedy
+     * @param  Integer $id
      * @return \Illuminate\Http\Response
      */
     public function update(Request $request, $id)
     {
-        $remedy = Remedy::find($id);
-        $remedy->update($request->all());
+        $remedy = $this->find($id);
+        if (!is_array($remedy)) {
+            $remedy->update($request->all());
+        }
         return $remedy;
     }
 
     /**
      * Remove the specified resource from storage.
      *
-     * @param  \App\Models\Remedy  $remedy
+     * @param  Integer $id
      * @return \Illuminate\Http\Response
      */
     public function destroy($id)
     {
-        $remedy = Remedy::find($id);
-        $remedy->delete();
-        return ['msg' => 'A medicação foi removida com sucesso'];
+        $remedy = $this->find($id);
+        if (!is_array($remedy)) {
+            $remedy->delete();
+            $remedy = ['msg' => 'A medicação foi excluída com sucesso'];
+        }
+        return $remedy;
     }
 }
